+++ CHANGELOG.md
@@ -12,3 +12,6 @@
 * **BC** - signature of `Psl\Type\object` function changed from `object<T of object>(classname<T> $classname): TypeInterface<T>` to `object(): TypeInterface<object>` ( to preserve the old behavior, use `Psl\Type\instance_of` )
 * introduced `Psl\Type\instance_of` function, with the signature of `instance_of<T of object>(classname<T> $classname): TypeInterface<T>`.
 * introduced a new `Psl\Async` component.
+* introduced a new `Psl\IO\Stream` component.
+* refactored `Psl\IO` handles API.
+* introduced a new `Psl\File` component.